+++ package.json
@@ -47,7 +47,7 @@
     "check:cspell": "cspell lint **/*.ts **/README.md .changeset/*.md --no-progress",
     "check:dep": "vsh check-dep",
     "check:type": "turbo run typecheck",
-    "clean": "node ./scripts/clean.mjs",
+    "clean": "node ./scripts/common/clean.mjs",
     "commit": "czg",
     "dev": "turbo-run dev",
     "dev:antd": "pnpm -F @arco/web-antd run dev",
