+++ scripts/turbo-run/src/run.ts
@@ -53,7 +53,7 @@ export async function run(options: RunOptions) {
     process.exit(1);
   }
 
-  execaCommand(`pnpm --filter=${String(selectPkg)} run ${command}`, {
+  await execaCommand(`pnpm --filter=${String(selectPkg)} run ${command}`, {
     stdio: 'inherit',
   });
 }
